+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network.WebrtcAdapter/WebrtcNetworkAdapter.ts
